// src/types/activity.ts
export interface Activity {
    activityType: string;
    date: string;
    favorite: boolean;
    title: string;
    distance: string;
    calories: string;
    time: string;
    avgHR: string;
    maxHR: string;
    avgRunCadence: string;
    maxRunCadence: string;
    avgPace: string;
    bestPace: string;
    totalAscent: string;
    totalDescent: string;
    avgStrideLength: string;
    trainingStressScore: string;
    steps: string;
    minTemp: string;
    decompression: string;
    bestLapTime: string;
    numberOfLaps: string;
    maxTemp: string;
    movingTime: string;
    elapsedTime: string;
    minElevation: string;
    maxElevation: string;
  }
  